drivers: spi: cc13xx_cc26xx: Allow clocks below 2MHz
Remove unnecessary check that the SPI clock is being set to a frequency
above 2MHz to allow devices running at common lower frequencies (i.e.
1.2MHz and 400kHz).

Replace with check that the frequency is not below the minimum frequency
supported by the chipset to prevent overflow error which can occur if
the HAL sets a frequency too low resulting in a SPI clock much larger
than expected.

Fixes #69986